+++ nbforager/__init__.py
@@ -1,5 +1,6 @@
 """nbforager."""
 
+from nbforager.foragers.joiner import Joiner
 from nbforager.nb_api import NbApi
 from nbforager.nb_forager import NbForager
 from nbforager.nb_tree import NbTree
@@ -8,10 +9,11 @@ from nbforager.parser.nb_parser import NbParser
 from nbforager.parser.nb_value import NbValue
 
 __all__ = [
+    "Joiner",
     "NbApi",
-    "NbParser",
     "NbCustom",
     "NbForager",
+    "NbParser",
     "NbTree",
     "NbValue",
 ]