+++ tests/foragers_/test__gardener.py
@@ -7,13 +7,13 @@ from nbforager.api.base_c import BaseC
 from nbforager.foragers.ipv4 import IPv4
 from nbforager.foragers.joiner import Joiner
 from nbforager.nb_tree import NbTree
-from tests import objects
+from tests import functions as func
 
 
 @pytest.fixture
 def joiner() -> Joiner:
     """Init Joiner with root data."""
-    tree: NbTree = objects.full_tree()
+    tree: NbTree = func.full_tree()
     tree = nb_tree.join_tree(tree)
     joiner_ = Joiner(tree=tree)
     joiner_.init_extra_keys()
@@ -53,6 +53,18 @@ def test__join_dcim_devices(joiner: Joiner):
     for key in reserved_keys:
         isinstance(interface_d[key], dict)
     assert interface_d["_ip_addresses"]["10.0.0.1/24"]["address"] == "10.0.0.1/24"
+
+
+@pytest.mark.parametrize("params, expected", [
+    ({}, [1, 2, 3]),
+    ({"id": 1}, [1]),
+    ({"device_role__name": "DEVICE ROLE1"}, [1, 2]),
+])
+def test__join_dcim_devices__filter(joiner: Joiner, params, expected):
+    """Joiner.join_dcim_devices() with filtering params."""
+    joiner.join_dcim_devices(**params)
+    actual = [i for i, d in joiner.tree.dcim.devices.items() if d["_interfaces"]]
+    assert actual == expected
 
 
 def test__join_virtualization_virtual_machines(joiner: Joiner):